Add stats command Add support for rename between different folders Add rename operation to ObjectStorage Fix open() when opening the same file multiple times, now can only be done with the readonly flag Fix stat TTL value Fix release() no cleaning up open_files Remove get_uniqueness_test() and add param ObjInfo to remove() in ObjectStorage Rename file_changes.file_sha512 to file_changes.file_hash Rename module sql to metadata_db Reduce logging in debug mode

+++ src/obj_storage/fs_object_storage.rs
@@ -3,12 +3,14 @@ use std::path::PathBuf;
 use std::rc::Rc;
 use anyhow::{anyhow, Context};
 use log::{error, info};
+use crate::AnyError;
 use crate::config::StorageConfig;
 use crate::obj_storage::{ObjInfo, ObjectStorage, UniquenessTest};
+use crate::storage::ObjInUseFn;
 
 pub struct FsObjectStorage {
     pub base_path: PathBuf,
-    pub config: Rc<StorageConfig>
+    pub config: Rc<StorageConfig>,
 }
 
 impl FsObjectStorage {
@@ -20,14 +22,14 @@ impl FsObjectStorage {
 }
 
 impl ObjectStorage for FsObjectStorage {
-    fn get(&mut self, info: &ObjInfo) -> Result<Vec<u8>, anyhow::Error> {
+    fn get(&mut self, info: &ObjInfo) -> Result<Vec<u8>, AnyError> {
         let path = self.path(&info);
         info!("Get: {:?}", &path);
 
         fs::read(&path).context("FS failed to read file")
     }
 
-    fn put(&mut self, info: &mut ObjInfo, content: &[u8]) -> Result<(), anyhow::Error> {
+    fn put(&mut self, info: &mut ObjInfo, content: &[u8]) -> Result<(), AnyError> {
         let path = self.path(&info);
         info!("Put: {:?}", &path);
 
@@ -37,8 +39,19 @@ impl ObjectStorage for FsObjectStorage {
         fs::write(&path, content).context("FS failed to write file")
     }
 
-    fn remove(&mut self, info: &ObjInfo) -> Result<(), anyhow::Error> {
+    fn remove(&mut self, info: &ObjInfo, is_in_use: ObjInUseFn) -> Result<(), AnyError> {
         let path = self.path(&info);
+        let test = if self.config.use_hash_as_filename {
+            UniquenessTest::Sha512
+        } else {
+            UniquenessTest::Path
+        };
+
+        // If is object in use by other file (deduplication), do not remove it
+        if is_in_use(info, test)? {
+            return Ok(());
+        }
+
         info!("Remove: {:?}", &path);
 
         fs::remove_file(&path).map_err(|e| {
@@ -46,7 +59,30 @@ impl ObjectStorage for FsObjectStorage {
         })
     }
 
-    fn nuke(&mut self) -> Result<(), anyhow::Error> {
+    fn rename(&mut self, prev_info: &ObjInfo, new_info: &ObjInfo) -> Result<(), AnyError> {
+        let prev_path = self.path(&prev_info);
+        let new_path = self.path(&new_info);
+
+        if prev_path == new_path {
+            return Ok(());
+        }
+
+        // Noop, previous file does not exist
+        if fs::metadata(&prev_path).is_err() {
+            return Ok(());
+        }
+
+        info!("Rename: {:?} -> {:?}", &prev_path, &new_path);
+
+        if let Some(parent) = new_path.parent() {
+            fs::create_dir_all(parent).context("FS failed to create dir")?;
+        }
+        fs::rename(&prev_path, &new_path).map_err(|e| {
+            anyhow!("FS failed to rename '{:?}' -> '{:?}': {:?}", prev_path, new_path, e)
+        })
+    }
+
+    fn nuke(&mut self) -> Result<(), AnyError> {
         info!("Nuke: {:?}", &self.base_path);
 
         for entry_res in fs::read_dir(&self.base_path)? {
@@ -72,13 +108,5 @@ impl ObjectStorage for FsObjectStorage {
         }
 
         Ok(())
-    }
-
-    fn get_uniqueness_test(&self) -> UniquenessTest {
-        if self.config.use_hash_as_filename {
-            UniquenessTest::Sha512
-        } else {
-            UniquenessTest::Path
-        }
     }
 }
